+++ mex/mex.py
@@ -114,6 +114,7 @@ app.add_page(
         IngestState.refresh,
         IngestState.resolve_identifiers,
         IngestState.flag_ingested_items,
+        IngestState.resolve_primary_source_titles,
     ],
 )
 app.add_page(
@@ -132,8 +133,6 @@ app.add_page(
     title="MEx Consent",
     on_load=[
         State.check_ldap_login,
-        ConsentState.get_all_data,
-        ConsentState.resolve_identifiers,
         ConsentState.get_consent,
     ],
 )