+++ tests/ingest/test_main.py
@@ -1,14 +1,9 @@
-from typing import TYPE_CHECKING
-
 import pytest
 from playwright.sync_api import Page, expect
 
 from mex.common.backend_api.connector import BackendApiConnector
-from mex.editor.ingest.models import ALL_AUX_PROVIDERS, AuxProvider
+from mex.editor.ingest.models import ALL_AUX_PROVIDERS, AuxProviderKey
 from tests.conftest import build_pagination_regex, build_ui_label_regex
-
-if TYPE_CHECKING:
-    from re import Pattern
 
 
 @pytest.fixture
@@ -140,23 +135,19 @@ def test_search_and_ingest_roundtrip(
 
 @pytest.mark.integration
 def test_infobox_visibility_and_content(ingest_page: Page) -> None:
-    expected_callout_content: dict[AuxProvider, Pattern[str]] = {
-        AuxProvider.LDAP: build_ui_label_regex("ingest.search_info.ldap"),
-        AuxProvider.WIKIDATA: build_ui_label_regex("ingest.search_info.wikidata"),
+    page = ingest_page
+    expected_callout_content = {
+        AuxProviderKey.LDAP: build_ui_label_regex("ingest.search_info.ldap"),
+        AuxProviderKey.WIKIDATA: build_ui_label_regex("ingest.search_info.wikidata"),
+        AuxProviderKey.ORCID: None,
     }
 
-    page = ingest_page
-
     for provider in ALL_AUX_PROVIDERS:
-        tab = page.get_by_role("tab", name=provider)
-        tab.click(
-            timeout=50_000
-        )  # high timeout cuz tab might be disabled due to loading
+        tab = page.get_by_role("tab", name=provider.dynamic_name)
+        tab.click()
+        callout = page.get_by_test_id("ingest-infobox-callout")
 
-        tab_content = page.locator(f"[id*='{provider}'][role='tabpanel']")
-        callout = tab_content.locator(".rt-CalloutRoot")
-
-        expected_content = expected_callout_content.get(provider)
+        expected_content = expected_callout_content[provider.key]
         if expected_content:
             expect(callout).to_have_count(1)
             expect(callout).to_have_text(expected_content)